Add `money / price` in place of `price * money`
To get a Quantity, the operation should be `money / price`, not
`price * money`.

E.g. when money is in USD and price is in USD/hour, the expression
`price * money` is supposed to have a dimension of `USD^2 / hour`, not
`hour`. To get the dimension `hour`, the expression `money / price`
needs to be used.