Create method to parse Currency from String in companion object
This adds new method in the Currency companion to parse a Currency from
a String given a MoneyContext. The method returns a Try[Currency], which
is a Failure when the string code is not found in the set of currencies
available in the MoneyContext.

It also makes all methods in the Money companion that used to parse the
currency symbol directly use instead the Currency companion and receive
an implicit MoneyContext.